ai/collect_data_google_custom_search_api/requirements.txt
astroid==2.9.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
google-api-core==2.5.0
google-api-python-client==1.12.10
google-auth==2.6.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.1
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
pip-autoremove==0.10.0
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pylint==2.12.2
pyparsing==3.0.7
python-decouple==3.6
python-resize-image==1.1.20
requests==2.27.1
rsa==4.8
six==1.16.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.8
windows-curses==2.3.0
wrapt==1.13.3
zipp==3.7.0